light restructuring
